refactor some param, move bruteforce to algorithm (#490)
- refactor hgraph search param
- move bruteforce to algorithm/
