+++ src/algorithm/brute_force.cpp
