+++ src/algorithm/hnswlib/hnswalg.h
@@ -55,7 +55,6 @@ struct CompareByFirst {
 using MaxHeap = std::priority_queue<std::pair<float, InnerIdType>,
                                     vsag::Vector<std::pair<float, InnerIdType>>,
                                     CompareByFirst>;
-const static float THRESHOLD_ERROR = 2e-6;
 
 class HierarchicalNSW : public AlgorithmInterface<float> {
 private:
