+++ src/quantization/scalar_quantization/fp16_quantizer.h
@@ -163,9 +163,8 @@ inline float
 FP16Quantizer<metric>::ComputeImpl(const uint8_t* codes1, const uint8_t* codes2) {
     if constexpr (metric == MetricType::METRIC_TYPE_L2SQR) {
         return FP16ComputeL2Sqr(codes1, codes2, this->dim_);
-    } else if constexpr (metric == MetricType::METRIC_TYPE_IP) {
-        return 1 - FP16ComputeIP(codes1, codes2, this->dim_);
-    } else if constexpr (metric == MetricType::METRIC_TYPE_COSINE) {
+    } else if constexpr (metric == MetricType::METRIC_TYPE_IP or
+                         metric == MetricType::METRIC_TYPE_COSINE) {
         return 1 - FP16ComputeIP(codes1, codes2, this->dim_);
     } else {
         return 0;
@@ -193,9 +192,8 @@ FP16Quantizer<metric>::ComputeDistImpl(Computer<FP16Quantizer>& computer,
     auto* buf = computer.buf_;
     if constexpr (metric == MetricType::METRIC_TYPE_L2SQR) {
         dists[0] = FP16ComputeL2Sqr(buf, codes, this->dim_);
-    } else if constexpr (metric == MetricType::METRIC_TYPE_IP) {
-        dists[0] = 1 - FP16ComputeIP(buf, codes, this->dim_);
-    } else if constexpr (metric == MetricType::METRIC_TYPE_COSINE) {
+    } else if constexpr (metric == MetricType::METRIC_TYPE_IP or
+                         metric == MetricType::METRIC_TYPE_COSINE) {
         dists[0] = 1 - FP16ComputeIP(buf, codes, this->dim_);
     } else {
         throw VsagException(ErrorType::INTERNAL_ERROR, "unsupported metric type");
